+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 
 # v1.0.0~beta3
 * Add in ``ifunny`` watermark template.
+* Add in ``america`` template.
+* Add in Minecraft achievement generator.
+* Add in ``piper`` template.
+* Templates which accept external images now accept ``webp`` and ``tiff`` images.
+* ``/api/*`` now has some form of dynamic routing.